[CICD-761] Removing make remote relative method
With the help of Michael, it was determined that the easiest way to make
sure that the excludes files is fully used for the rsynch command, is to
make sure to use the file name, but not the full file path. This will
make sure the files are excluded, no matter which folder location you
attempt to deploy to. So default base of the wordpress install will
work, or if you use within wp-content or other folder locations.

Removing out the testing calls for now too, until we can get a better
mocking framework in place to make sure synch_files can be full tested.

Co-authored-By: Michael Day <michael.day@wpengine.com>