[CICD-761] Generate dynamic excludes for rsync (#42)
* [CICD-761] Explicitly export color globals

Resolves a warning about unused variables in the common file.

* [CICD-761] Create tests/helpers dir

Also moves common.sh to new utils dir.

* [CICD-761] Create print_mount_paths util

* [CICD-761] Generate excludes list dynamically

Outputs the list to stdout, which is called with process substitution in the initial rsync command.

Also disables debug mode for the block calling the generate function. This prevents unwanted debug output
(specific to the excludes list generation) from being shown.

* [CICD-761] Move to a simpler approach for unit tests

* [CICD-761] Bump version with patch update

* [CICD-761] Glob the current dir instead of checking  mount paths

Simplifies the logic for determining what files were copied to workdir by globbing for certain dirs instead of parsing proc/self/mountinfo.

This is also better for other mounted dir structures (such as when a GHA runner mounts the image); we are simply checking what is in the current dir instead of inferring the structure based off mounted source path.

* [CICD-761] Use only REMOTE_PATH for generating excludes

The goal of the excludes list is to prevent users from accidentally
breaking their site by overwriting or deleting important files on
the server.

While there are many files that could be deemed important, we explicily
want to exclude a handful of files from the site root, some from wp-content,
and some from wp-content/mu-plugins.

In order to determine if a deploy is at risk of overwriting or deleting an
important file on the server, we just need to look at the REMOTE_PATH.

If REMOTE_PATH is unset, we know the destination is the site root.
If REMOTE_PATH ends in wp-content, we know the destination is wp-content.
If REMOTE_PATH ends in wp-content/mu-plugins, we know the destination is
 wp-content/mu-plugins.

All three of these cases carry the risk of overwritting or deleting
important files. In order to prevent this, we need to make the exclude
rules relative to the REMOTE_PATH.

SRC_PATH no longer matters because rsync assumes it to be identical to
REMOTE_PATH for the sake of determining transfer/deletion deltas.

* [CICD-761] Add additional cases for generated exclude from rules

Tests how different REMOTE_PATHS affect the generated exclude
rule output.

* [CICD-761] nit: update verbiage for batch tests success message

---------

Co-authored-by: Andrew Matthews <andrew.matthews@wpengine.com>