Remove notify step from e2e deploy (#44)
Removes the notify step of our e2e workflow. The failures in this workflow have consistently been unrelated to the site-deploy image itself and are therefore not actionable. Instead, they indicate an issue with some other service we depend on (GitHub, Docker Hub, SSH Gateway, WPE User Portal). These services are either out of our control or are already monitored in other ways.

This also bumps the cron schedule to 1 per hour.

Bonus: gets rid of a dependency on voxmedia/github-action-slack-notify-build, which has been deprecated.