Remove dependency on instrumentisto/rsync-ssh
Dependabot version checks aren't able to detect changes to the base
image because the tag format (i.e. alpine3.20) is not standard semver.
This prevents Dependabot from automatically updating the base image when
a new version is released, creating the need to manually monitor for and
apply base image updates.

The instrumentisto/rsync-ssh base image isn't particularly complex and
we're already running package updates and installing a few additional
dependencies. Therefore, it makes sense to just use alpine directly and
install all of the instrumentisto/rsync-ssh dependencies ourselves. This
should allow Dependabot to automatically update the base image when a
new version is released.